requirements.txt
numpy
PyYAML
optuna
jax
matplotlib
scikit-learn
plotly
optax
